XMLHttpRequest/open-method-responsetype-set-sync.htm

<!DOCTYPE html>
<html>
  <head>
    <title>XMLHttpRequest: open() sync request not allowed if responseType is set</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="help" href="http://dvcs.w3.org/hg/xhr/raw-file/tip/Overview.html#the-open()-method" data-tested-assertations="following::ol[1]/li[4]" />

  </head>
  <body>
    <div id="log"></div>
    <script>
      function request(type) {
        test(function() {
          var client = new XMLHttpRequest()
          client.responseType = type
          assert_throws("InvalidAccessError", function() { client.open('GET', "...", false) })
        }, document.title + " (" + type + ")")
      }
      request("arraybuffer")
      request("blob")
      request("json")
      request("text")
      request("document")
    </script>
  </body>
</html>
